Added victory spectator mode, user can view players cards when they exit the game (until the end of the game).
fix: potential race condition at the victory, the game tries update the state at the end of the round despite it not having the proper state.

potential issue (under investigation): there appears to be a mem leak between component state between previous round and current round, I have avverted in (this.handleGameStateUpdate). q: Is it unsafe to access state in handleGameStateUpdate?